+++ src/index.ts
@@ -1,9 +1,11 @@
 import transfer from './transfer';
 import billPayment from './bill-payment';
+import pix from './pix';
 
 const occurrence = Object.freeze({
   transfer,
-  billPayment
+  billPayment,
+  pix
 });
 
 export default occurrence;