+++ src/pix/BaasPixCode.ts
@@ -1,0 +1,26 @@
+export enum BaasPixCode {
+  /**
+   * Registro criado
+   */
+  PIX0001 = "PIX0001",
+  /**
+   * Confirmado
+   */
+  PIX0002 = "PIX0002",
+  /**
+   * Não confirmado
+   */
+  PIX0003 = "PIX0003",
+  /**
+   * Rejeitado
+   */
+  PIX0004 = "PIX0004",
+  /**
+   * Realizado
+   */
+  PIX0005 = "PIX0005",
+  /**
+   * Cancelado
+   */
+  PIX0006 = "PIX0006",
+}