+++ src/pix/index.ts
@@ -1,0 +1,5 @@
+import {BaasPixCode} from './BaasPixCode';
+
+export default Object.freeze({
+  ...BaasPixCode
+});